+++ app/src/main/java/com/minizuure/todoplannereducationedition/services/database/task/TaskTableDao.kt
@@ -16,13 +16,13 @@ interface TaskTableDao : BaseIODao<TaskTable> {
 
 
     @Query("SELECT * FROM task_table WHERE id = :id")
-    suspend fun getById(id: Int): TaskTable?
+    suspend fun getById(id: Long): TaskTable?
 
     @Query("SELECT * FROM task_table WHERE index_day = :indexDay")
     suspend fun getByIndexDay(indexDay: Int): List<TaskTable>
 
     @Query("SELECT * FROM task_table WHERE is_custom_session = 1 AND session_id = :sessionId")
-    suspend fun getBySessionId(sessionId: Int): List<TaskTable>
+    suspend fun getBySessionId(sessionId: Long): List<TaskTable>
 
 
 }
